requirements.txt
# Core dependencies
requests>=2.31.0
discord-webhook>=1.3.0

# Web scraping (for fallback APIs)
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Data handling
python-dotenv>=1.0.0

# Optional: Better scoring
aiohttp>=3.9.0  # For async requests if needed
